==> FINAL/scripts/01_intro.qmd <==
---
title: "Chemical Sample Classification Report"
subtitle: "Introduction"
author: "Petter Lövehagen"
date: "08 May 2023"

RVersion: 3.6.0 
RStudio: 2023.03.0+386 Cherry Blossom Release
Platform: "x86_64-w64-mingw32/x64 (64-bit)"
OS: "Windows 11 SE"

format: 
  docx: default
  html:
    toc: true
    toc-depth: 2
    toc-title: "Contents"
    toc-location: left
    code-fold: true
    code-line-numbers: true
    theme: spacelab
    html-math-method: mathjax
    highlight: "tango"
    #css: "../data/custom.css"
prefer-docx: true
lang: "en-GB"
subject: "chemical sample classifition"
bibliography: "../data/AS_CW_References.bib"
csl: "../data/harvard-university-of-the-west-of-england.csl"

knitr:
  opts_chunk:
    include: false
    tidy: true
    echo: false
    warning: false
    message: false
    error: true
    fig.align: left
    out.width: '90%'
    fig.width: 6
    fig.height: 4
---

## Brief {#sec-brief}

The brief is to develop a `classification tool` to classify chemicals into classes.

In addition, the client is interested in understanding whether all variables are *necessary* for achieving optimal out-of-sample classification performance, or if a subset of measurements will suffice.

This report will aim to:

-   explore, analyse and feedback on the provided dataset
-   develop an optimised, classification model
-   comment on whether all variables are required
-   provide recommendations and options, as applicable

::: callout-note
Additional technical details are available in the code comments.
:::
