==> FINAL/scripts/04_dim_red.qmd <==
---
title: "Chemical Sample Classification Report"
subtitle: "Dimension Reduction"
author: "Petter Lövehagen"
date: "08 May 2023"

RVersion: 3.6.0 
RStudio: 2023.03.0+386 Cherry Blossom Release
Platform: "x86_64-w64-mingw32/x64 (64-bit)"
OS: "Windows 11 SE"

format: 
  docx: default
  html:
    toc: true
    toc-depth: 3
    toc-title: "Contents"
    toc-location: left
    code-fold: true
    code-line-numbers: true
    theme: spacelab
    html-math-method: mathjax
    highlight: "tango"
    #css: "../data/custom.css"
prefer-docx: true
lang: "en-GB"
bibliography: "../data/AS_CW_References.bib"
csl: "../data/harvard-university-of-the-west-of-england.csl"

knitr:
  opts_chunk:
    include: false
    tidy: true
    echo: false
    warning: false
    message: false
    error: true
    fig.align: 'left'
    out.width: '90%'
    # fig.width: 6
    # fig.height: 4
---

<!--# The 'chem_packages.R file installs/loads required packages for this section, sets a seed (for reproducibility) and loads flextable defaults. Check for further details. -->

```{r dim_red_packages}
# load/install R packages/libraries
# set seed
# load flextable defaults

file <- "dim_red"
source("chem_packages.R")


```

## load scaled data with and without labels

dimension reduction:

PCA - identify principal components, linear combos of original variabes explaining the most variation in the data, plot to see in a lower dimension space t-SNE

PCA:

like a plot with PC1, PC2 the actual 5 classes (shape and label by label) pca arrows and variable labels

(see penguin example)

```{r}
KMO(train[1:20])
```

OVeral KMO is 0.71 which is above 0.5 indicating that the dataset has sufficient correlation present to warrant PCA/FA. The variable with the lowest ?? is X15 with 0.46.

some citation here.

parallel analysis [citation]

```{r}
scree(num_train)
```

```{r}
paran(num_train, iterations = 5000)
```
```{r}
library(ade4)
library(factoextra)
pca <- dudi.pca(num_train,scannf = F, nf=4)
fviz_pca_biplot(pca)
fviz_pca_biplot(pca, label="var",title="Chemical variables", repel=TRUE)
fviz_screeplot(pca)
get_eigenvalue(pca)
fviz_pca_biplot(pca,repel=T,label="var",axes=c(1,2))
fviz_pca_biplot(pca,repel=T,label="var",axes=c(1,3))
```

```{r}

## write your code here
fviz_pca_biplot(pca,
                repel=T,
                label="var",
                col.ind = train$label)
fviz_pca_biplot(pca,axes=c(1,3),
                repel=T,
                label="var",
                col.ind = train$label)


fviz_pca_ind(pca,
             geom.ind = "point", # show points only (but not "text")
             col.ind = train$label, # colour by groups
             addEllipses = TRUE, # Concentration ellipses
            legend.title = "Chemicals"
            )

fviz_pca_ind(pca,axes=c(1,3),
             geom.ind = "point", # show points only (but not "text")
             col.ind = train$label, # colour by groups
             addEllipses = TRUE, # Concentration ellipses
            legend.title = "Chemicals"
            )
```
